Refactored AD related entity queries
