+++ cmd/api/src/api/v2/ad_related_entity_test.go
@@ -1,17 +1,17 @@
 // Copyright 2023 Specter Ops, Inc.
-// 
+//
 // Licensed under the Apache License, Version 2.0
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
-// 
+//
 //     http://www.apache.org/licenses/LICENSE-2.0
-// 
+//
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-// 
+//
 // SPDX-License-Identifier: Apache-2.0
 
 package v2_test
@@ -20,6 +20,8 @@ import (
 	"net/http"
 	"testing"
 
+	"github.com/specterops/bloodhound/dawgs/ops"
+	"github.com/specterops/bloodhound/errors"
 	v2 "github.com/specterops/bloodhound/src/api/v2"
 	"github.com/specterops/bloodhound/src/api/v2/apitest"
 	dbMocks "github.com/specterops/bloodhound/src/database/mocks"
@@ -27,8 +29,6 @@ import (
 	"github.com/specterops/bloodhound/src/queries"
 	graphMocks "github.com/specterops/bloodhound/src/queries/mocks"
 	"go.uber.org/mock/gomock"
-	"github.com/specterops/bloodhound/dawgs/ops"
-	"github.com/specterops/bloodhound/errors"
 )
 
 func setup(t *testing.T) (*gomock.Controller, *graphMocks.MockGraph, *dbMocks.MockDatabase, v2.Resources) {
